requirements.txt
# Core Azure and AI Libraries
azure-identity>=1.14.0                # Entra ID authentication (DefaultAzureCredential)
azure-ai-evaluation>=0.2.0            # QAEvaluator, ContentSafetyEvaluator
openai>=1.0.0                         # OpenAI SDK (includes AzureOpenAI client)
azure-core>=1.28.0                    # Azure SDK core utilities

# Web and Data Processing
requests>=2.31.0                      # HTTP requests for fetching SharePoint pages
beautifulsoup4>=4.12.0                # HTML parsing and cleaning
scikit-learn>=1.3.0                   # TF-IDF for simple document scoring

# Testing and Development
pytest>=7.4.0                         # Unit testing framework
pytest-cov>=4.1.0                     # Coverage reporting
pytest-asyncio>=0.21.0                # Async test support

# Configuration
python-dotenv>=1.0.0                  # .env file loading

# Data and Serialization
pydantic>=2.0.0                       # Data validation and models
